+++ interface/src/components/WalletList.vue
@@ -51,12 +51,12 @@
     >
       <a-row>
         <a-col :span="8">
-          <a-form-item label="Nickname: " ref="nick" name="nick">
+          <a-form-item label="Nickname: " ref="nick" name="nick" v-bind="validateInfos.nick">
             <a-input v-model:value="formState.nick" placeholder="UnBankedKing" />
           </a-form-item>
         </a-col>
         <a-col :span="8">
-          <a-form-item label="Address: " ref="address" name="address">
+          <a-form-item label="Address: " ref="address" name="address" v-bind="validateInfos.address">
             <a-input
               v-model:value="formState.address"
               placeholder="0xeeee111122223333444455556666777788889999"
@@ -64,8 +64,8 @@
           </a-form-item>
         </a-col>
         <a-col :span="8">
-          <a-form-item label="Tags: " ref="tags" name="tags">
-            <a-input v-model:value="formState.tags" placeholder="abc one-two three" />
+          <a-form-item label="Tags: " ref="tags" name="tags" v-bind="validateInfos.tags">
+            <a-input v-model:value="formState.tags" placeholder="abc, one-two, three four" />
           </a-form-item>
         </a-col>
       </a-row>
@@ -75,7 +75,7 @@
           <a-button
             type="primary"
             class="bg-slate-600"
-            @click="onSubmit"
+            @click.prevent="onSubmit"
             :loading="awaitingNewWallet"
             :disabled="awaitingNewWallet"
           >
@@ -92,11 +92,12 @@ import Immutable from 'immutable';
 import WalletTagEdit from '@/components/WalletTagEdit.vue';
 import { CheckOutlined, EditOutlined } from '@ant-design/icons-vue';
 import { cloneDeep } from 'lodash-es';
-import { pushWallet, pullWallet, pushName } from '@/api/books.ts';
+import { pushWallet, pullWallet, pushName } from '@/api/books';
 import { computed, defineComponent, reactive, ref, toRaw } from 'vue';
-import { mapState, useStore } from 'vuex';
-import type { PropType } from 'vue';
-import { Address, WalletDetails } from '@/types';
+import { useStore } from 'vuex';
+import { WalletDetails } from '@/types';
+import { Form } from 'ant-design-vue';
+import { arrayAndHepTags } from '@/api/books';
 
 export default defineComponent({
   setup() {
@@ -109,11 +110,6 @@ export default defineComponent({
     //  mapState and mapGetters replacements
     const myWallets = computed(() => store.state.books.myWallets);
     const myFriends = computed(() => store.state.books.myFriends);
-    const namesInUse = computed(() => {
-      const myNames = this.myWallets.map((item) => item[1].nick);
-      const urNames = this.myFriends.map((item) => item[1].nick);
-      return myNames.concat(urNames);
-    });
 
     const count = computed(() => wallets.value.length + 1);
 
@@ -184,7 +180,6 @@ export default defineComponent({
     // Refs
     const overallLoading = ref(false);
     const formRef = ref();
-    const inputRef = ref();
     const awaitingNewWallet = ref(false);
 
     //  handlers
@@ -258,68 +253,68 @@ export default defineComponent({
       nick: [
         {
           required: true,
-          message: 'Nickname Required',
-          trigger: 'blur',
+          message: 'Nickname required',
         },
         {
           min: 2,
-          message: 'Nickname Required',
-          trigger: 'blur',
+          message: 'Nickname must be longer',
         },
       ],
       address: [
         {
           required: true,
-          message: 'Address Required',
-          trigger: 'blur',
+          message: 'Address required',
         },
         {
           min: 42,
           max: 42,
           pattern: /^0x[a-fA-F0-9]{40}$/,
-          message: 'Provide Wallet Address',
-          trigger: 'blur',
+          message: 'Provide valid wallet address',
         },
       ],
       tags: [
         {
           required: false,
-          pattern: /^[a-zA-Z0-9\-\_\s]+$/,
-          trigger: 'blur',
-          message: "a-z, 0-9, '-' and '_' only, separated by spaces",
+          pattern: /^[a-zA-Z0-9\-\s\,]*$/,
+          message: "a-z, 0-9, '-' and spaces only, separated by commas",
         },
       ],
     };
 
+    const useForm = Form.useForm
+    const { resetFields, validate, validateInfos } = useForm(formState, rules, {
+      onValidate: (...args) => console.log(...args),
+    });
+
     //  methods
     const onSubmit = () => {
-      awaitingNewWallet.value = true;
-      overallLoading.value = true;
-      formRef.value
-        .validate()
-        .then(() => {
-          console.log(formRef.value);
-          console.log('values', formState, toRaw(formState));
-          pushWallet(
-            toRaw(formState).address,
-            toRaw(formState).nick,
-            toRaw(formState).tags
-          )
-            .then((r) => {
-              console.log('res: ', r);
-            })
-            .catch((e) => {
-              console.log('err: ', e);
-            })
-            .finally(() => {
-              overallLoading.value = false;
-              awaitingNewWallet.value = false;
-              formRef.value.resetFields();
-            });
-        })
+      validate()
+      .then(() => {
+        console.log('values', formState, toRaw(formState));
+        awaitingNewWallet.value = true;
+        overallLoading.value = true;
+        const safeTags = arrayAndHepTags(toRaw(formState).tags)
+        pushWallet(
+          toRaw(formState).address,
+          toRaw(formState).nick,
+          safeTags
+        )
+          .then((r) => {
+            console.log('res: ', r);
+          })
+          .catch((e) => {
+            console.log('err: ', e);
+          })
+          .finally(() => {
+            overallLoading.value = false;
+            awaitingNewWallet.value = false;
+            resetFields();
+          });
+      })
       .catch((error) => {
+        // validation failed
         console.log('error', error);
-      });
+      })
     };
 
     return {
@@ -337,11 +332,11 @@ export default defineComponent({
       wallets,
       handleClose,
       handleInput,
-      namesInUse,
       allTags,
       columns,
       overallLoading,
       awaitingNewWallet,
+      validateInfos,
     };
   },