+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+
+- Validation for tags everywhere
+
+### Changed
+
+- Add loading state to tag manipulation on existing records
+- All API calls expect tag arrays; include helper function for handling
+- Disallow underscores in tags
+
+### Fixed
+
+- Validation on new Wallet, new Friend
+
 ## [1.1.0] - 2022-08-04
 
 ### Changed